requirements.txt
discord.py
psycopg2
requests
sqlalchemy == 1.3.19
git+https://github.com/Rapptz/discord-ext-menus